+++ VulkanMemoryAllocator/changelog.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## WIP
+- Bump VMA, no functional change
+- Use allocaBytes over allocaBytesAligned where possible
 
 ## [0.6] - 2021-06-22
 - Bump VMA, adding alignment info to PoolCreateInfo